+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Add a `regions` fn that returns a map of AWS regions and information
+  such as the name, domain, location, and partition of each region.
+
 ## v3.0.0 (2025-04-17)
 
 - (breaking) Remove invalid-keyword? function.